Was working on the problem of checking if a tree is balanced and got side tracked with trying to figure out how to display a tree in text.
Ended up having to figure out how to get a height and the max length of
the values on my way to getting the display to work.  I think I'm almost
there.  I need to figure out how to determine how much space is between
each node value when I have the tree height, the max node value length and
the current level of the tree I'm on.

I'm pretty sure I can display it by using a deque to get the values of
each level of the tree and putting it in the correct spot on an array of
possible locations and printing either the number of spaces of the max
node value length or printing the node value padded with spaces to match
the max node value length.

I also created some modules so I can share code with other tree problems I
will work on in the future.