Finished displayTrees().  I figured out that if the height is greater than 6, then it doesn't display that well, so I'm limiting the function to only display the first 6 levels of any tree that's passed in.
I also finished a function that detects if a tree is balanced.  Actually,
now that I think of it, it doesn't work.  I'm checking if a node has any
child that is None when I need to check that the node has both children
are None.  I'll go back to that in a sec.

I finished a function that takes a string of integers that are in order
and puts them into a tree in the correct location.