implemented a depth first search.  Used recursion to go through the children.  Since I'm using recursion and want to find a path, I need to pass the path in the function call (path is defaulted to empty list).  The escape case is if the start is equal to the end.  Then for each linked node, check if the node is not in the path.  If it is not, assign the return value to newPath.  If something is returned to newPath, return newPath
Implemented a breadth first search, which is pretty much like a level
traversal of a binary tree.  Use a queue and append linked nodes to the
end while popping nodes off the front a traversing.  Since graphs nodes
can link to any node, keep an dict of keys to figure out if nodes have
been visited.

I also implemented a function to find a path using breadth first search.
For this, I create a queue of paths.  while the queue is not empty, pop
off the top element, figure out the next nodes and append the updated
paths to the queue if they are not in the path already.  If they are, then
do nothing.